:recycle: chore: Try Deno 2

+++ boot/boot.sh
@@ -9,6 +9,7 @@ set -e
 PLATFORM=$(uname -ms)
 
 BUN_VERSION="1.0.30"
+DENO_VERSION="2.0.0"
 TAILWIND_VERSION="v3.1.6"
 
 # --- Internal ---
@@ -20,11 +21,14 @@ function download_tailwind {
     file=$dir/tailwind-$TAILWIND_VERSION
 
     case $PLATFORM in
-      'Darwin arm64') curl -fLo $file "https://github.com/tailwindlabs/tailwindcss/releases/download/$TAILWIND_VERSION/tailwindcss-macos-arm64";;
-      'Darwin x86_64') curl -fLo $file "https://github.com/tailwindlabs/tailwindcss/releases/download/$TAILWIND_VERSION/tailwindcss-macos-x64";;
-      'Linux arm64' | 'Linux aarch64') curl -fLo $file "https://github.com/tailwindlabs/tailwindcss/releases/download/$TAILWIND_VERSION/tailwindcss-linux-arm64";;
-      'Linux x86_64') curl -fLo $file "https://github.com/tailwindlabs/tailwindcss/releases/download/$TAILWIND_VERSION/tailwindcss-linux-x64";;
-      *) echo "Error: Unsupported platform: $PLATFORM"; exit 1;;
+    'Darwin arm64') curl -fLo $file "https://github.com/tailwindlabs/tailwindcss/releases/download/$TAILWIND_VERSION/tailwindcss-macos-arm64" ;;
+    'Darwin x86_64') curl -fLo $file "https://github.com/tailwindlabs/tailwindcss/releases/download/$TAILWIND_VERSION/tailwindcss-macos-x64" ;;
+    'Linux arm64' | 'Linux aarch64') curl -fLo $file "https://github.com/tailwindlabs/tailwindcss/releases/download/$TAILWIND_VERSION/tailwindcss-linux-arm64" ;;
+    'Linux x86_64') curl -fLo $file "https://github.com/tailwindlabs/tailwindcss/releases/download/$TAILWIND_VERSION/tailwindcss-linux-x64" ;;
+    *)
+      echo "Error: Unsupported platform: $PLATFORM"
+      exit 1
+      ;;
     esac
 
     mv $dir/tailwind-$TAILWIND_VERSION opt/tailwind
@@ -36,20 +40,22 @@ function download_tailwind {
 function download_bun {
   if [ ! -f ./opt/bun ]; then
     if ! command -v unzip >/dev/null; then
-      echo "Error: unzip is required to proceed"; exit 1
+      echo "Error: unzip is required to proceed"
+      exit 1
     fi
 
     dir=$(mktemp -d)
     zip=$dir/bun-$BUN_VERSION.zip
 
-    
-
     case $PLATFORM in
-      'Darwin arm64') NESTED_DIR="bun-darwin-aarch64";;
-      'Darwin x86_64') NESTED_DIR="bun-darwin-x64";;
-      'Linux arm64' | 'Linux aarch64') NESTED_DIR="bun-linux-x64";;
-      'Linux x86_64') NESTED_DIR="bun-linux-x64-baseline";;
-      *) echo "Error: Unsupported platform: $PLATFORM"; exit 1;;
+    'Darwin arm64') NESTED_DIR="bun-darwin-aarch64" ;;
+    'Darwin x86_64') NESTED_DIR="bun-darwin-x64" ;;
+    'Linux arm64' | 'Linux aarch64') NESTED_DIR="bun-linux-x64" ;;
+    'Linux x86_64') NESTED_DIR="bun-linux-x64-baseline" ;;
+    *)
+      echo "Error: Unsupported platform: $PLATFORM"
+      exit 1
+      ;;
     esac
 
     curl -fLo "$zip" "https://github.com/oven-sh/bun/releases/download/bun-v$BUN_VERSION/$NESTED_DIR.zip"
@@ -58,6 +64,37 @@ function download_bun {
     mv opt/$NESTED_DIR/bun opt/bun
     rm -rf opt/$NESTED_DIR
     chmod +x opt/bun
+    rm $zip
+  fi
+}
+
+function download_deno {
+  if [ ! -f ./opt/deno ]; then
+    if ! command -v unzip >/dev/null; then
+      echo "Error: unzip is required to proceed"
+      exit 1
+    fi
+
+    dir=$(mktemp -d)
+    zip=$dir/bun-$DENO_VERSION.zip
+
+    case $PLATFORM in
+    'Darwin arm64') NESTED_DIR="deno-aarch64-apple-darwin" ;;
+    'Darwin x86_64') NESTED_DIR="deno-x86_64-apple-darwin" ;;
+    'Linux arm64' | 'Linux aarch64') NESTED_DIR="deno-aarch64-unknown-linux-gnu" ;;
+    'Linux x86_64') NESTED_DIR="deno-x86_64-unknown-linux-gnu" ;;
+    *)
+      echo "Error: Unsupported platform: $PLATFORM"
+      exit 1
+      ;;
+    esac
+
+    curl -fLo "$zip" "https://github.com/denoland/deno/releases/download/v$DENO_VERSION/$NESTED_DIR.zip"
+
+    unzip -o $zip -d opt &>/dev/null
+    mv opt/$NESTED_DIR/deno opt/deno
+    rm -rf opt/$NESTED_DIR
+    chmod +x opt/deno
     rm $zip
   fi
 }
@@ -89,11 +126,11 @@ if [ ! -d ./bin ]; then
   mkdir ./bin
 fi
 
-## Download TailwindCSS
 download_tailwind
 
-## Download Bun
 download_bun
+
+download_deno
 
 ## Install bin dependencies
 opt/bun i
